# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MY API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3005",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/j": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Fxxx": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/bisection": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security":[
          {
            "bearerAuth":[]
          }
        ]
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Fxxx": {
                  "example": "any"
                },
                "Xl": {
                  "example": "any"
                },
                "Xr": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/onepoint": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gettoken": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "secureityDifinations": {
    "basic": {
      "type": "basic"
    },
    "apiKey": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}